Add Naive Bayes algorithm.
This algorithm will help to classify if a given team will lose or win.
The initial approach is to devide the dota_metrics.db in 70% of its matches
to learn and the rest to the algorithm perfomance.

The analytics app is in its early stages, I plan to add more
AI algorithms and different model approaches for each one of them.